+++ src/main/java/com/yanny/ytech/ModBusSubscriber.java
@@ -2,8 +2,10 @@ package com.yanny.ytech;
 
 import com.yanny.ytech.compatibility.TopCompatibility;
 import com.yanny.ytech.configuration.SpearType;
+import com.yanny.ytech.configuration.block_entity.WellPulleyBlockEntity;
 import com.yanny.ytech.configuration.entity.*;
 import com.yanny.ytech.configuration.item.BasketItem;
+import com.yanny.ytech.configuration.item.DiviningRodItem;
 import com.yanny.ytech.configuration.item.SpearItem;
 import com.yanny.ytech.configuration.model.*;
 import com.yanny.ytech.configuration.renderer.*;
@@ -36,7 +38,6 @@ import net.minecraftforge.event.entity.EntityAttributeCreationEvent;
 import net.minecraftforge.event.entity.SpawnPlacementRegisterEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.event.config.ModConfigEvent;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
 import org.jetbrains.annotations.NotNull;
@@ -59,6 +60,8 @@ public class ModBusSubscriber {
         event.enqueueWork(() -> {
             ItemProperties.register(YTechItems.BASKET.get(), BasketItem.FILLED_PREDICATE,
                     (stack, level, entity, seed) -> BasketItem.getFullnessDisplay(stack));
+            ItemProperties.register(YTechItems.DIVINING_ROD.get(), DiviningRodItem.ABUNDANCE_PREDICATE,
+                    (stack, level, entity, seed) -> entity != null ? (float) WellPulleyBlockEntity.getWaterAbundance(entity.getOnPos()) : 1);
             YTechItems.SPEARS.values().forEach((item) -> ItemProperties.register(item.get(), SpearItem.THROWING_PREDICATE,
                     (stack, level, entity, seed) -> entity != null && entity.isUsingItem() && entity.getUseItem() == stack ? 1.0F : 0.0F));