+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Crops need hydrated farmland to grow
 - Farmland has chance to consume water source
 - Added Well Pulley
+- Added Divining Rod for detecting water abundance for Well Pulley
 
 ## [0.8.0]